+++ android/app/src/main/java/app/priceguard/ui/home/theme/ThemeDialogFragment.kt
@@ -1,4 +1,4 @@
-package app.priceguard.ui.home
+package app.priceguard.ui.home.theme
 
 import android.app.Dialog
 import android.content.Context
