+++ android/app/src/main/res/navigation/nav_graph_home.xml
@@ -35,7 +35,7 @@
         app:destination="@id/myPageFragment" />
     <dialog
         android:id="@+id/themeDialogFragment"
-        android:name="app.priceguard.ui.home.ThemeDialogFragment"
+        android:name="app.priceguard.ui.home.theme.ThemeDialogFragment"
         android:label="ThemeDialogFragment"
         tools:layout="@layout/fragment_theme_dialog"/>
 </navigation>
